+++ client/Targets/Domain/Entities/Sources/My/Profile.swift
@@ -14,7 +14,7 @@ public struct Profile {
     public let userName: String
     public let profileImageURL: String?
     public let isFollow: Bool
-    public let temperature: Double
+    public let temperature: Int
     public let temperatureFeeling: String
     public let followerCount: Int
     public let storyCount: Int
@@ -28,7 +28,7 @@ public struct Profile {
         userName: String,
         profileImageURL: String?,
         isFollow: Bool, 
-        temperature: Double,
+        temperature: Int,
         temperatureFeeling: String,
         followerCount: Int,
         storyCount: Int,
